integrated location data source with api

+++ data/src/main/java/com/android/xu/data/entities/WeatherForecast.kt
@@ -1,4 +1,4 @@
 package com.android.xu.data.entities
 
-data class WeatherForecast(val currentTemp_c : Int, val cityName : String, val days : ArrayList<Forecastday>) {
+data class WeatherForecast(val currentTemp_c : Double, val cityName : String, val days : ArrayList<Forecastday>) {
 }
